Fluid Placeholder Block (#234)
Adds a new type of substitution/placeholder block: the Fluid Placeholder
Ensures that the destination block is solid or a fluid source
If the destination block is not a fluid source already, the default fluid type for the dimension is used (i.e. lava in nether)
renders in the preview as the default fluid type for that dimension
some example use cases:

almost all underwater structures should use this over a regular Placeholder block (this is part of the setup for underwater styles)
around the edge of ship schematics so that they don't take chunks out of the shore when placed close to land.
any decorations that work with the respective dimension fluid, for example a fountain that is a lava fountain in the nether
Requires a consecutive PR to minecolonies but works already when stand-alone.