Don't overwrite existing schematic names
Currently, the process of running a scan over an area that contains a
DecorationController from minecolonies will overwrite the schematicName
field in the DecorationController with the filename provided in the scan
tool.  This change modifies the process so that DecorationControllers
with existing schematicNames preserve their values, but empty
DecorationControllers will inherit the filename.