Feature/shape tool integration (#435)
Adds rotation/mirroring indicators recently added to build tool also to shape tool
Makes some preparatory changes to allow the shape tool to be integrated with MineColonies
One side effect of this is a paste button similar to the build tool, but in pure Structurize this behaves identically to the tick.
Another side effect is a new chat message shown if you try to use the tool in survival mode.
This also fixes a nasty bug in the way that the shape tool (or anything other than the scan tool) generated blueprints.
Adds +/- buttons to the shape dimensions for ease of use.
Fixes the </> buttons to the shape type that already existed but did nothing.
Removes some "todo remove if not used" code that was not used.
Hides the preview after pasting a shape (rather than leaving it), but allows the same shape to be quickly recalled in another location.