Feature/groundlevel (#436)
Adds a "Tag Anchor" block, which acts as an anchor block and can be used to hold Tag Tool tags when MineColonies is not loaded, or when making a decoration without a MineColonies Decoration Controller. On fancy placement it will always convert to air (so it can only be used for temporary tags, but is safe to rescan provided you only use the left paste).
Modifies the big_well default schematic to include a tag anchor and groundlevel tag.
Also removes the MineColonies waypoint that it previously erroneously had, and was auto-upgraded to the new shingles, plus a few other side-effects of rescanning.
Note that currently the tag anchor is slightly off-center, and so the well will rotate around that rather than around the blueprint center. I don't consider this a problem, but it could easily be fixed if you think it is.
Modifies the Build Tool UI so that it automatically offsets from the current anchor position to the ground level of the active schematic (such that as you switch between schematics it will always correctly align the ground level -- assuming that the player originally clicked on a block in the ground, as is usually the case when placing buildings).
If the player explicitly moves the preview up or down, then it will maintain that same relative placement as you switch between schematics.