Fix groundlevel again (#443)
Turns out Structurize and MineColonies used different (and inconsistent) methods to choose the initial block to anchor the build (and shape) tool previews at. (Partly because they have to hook different events to do so.) This changes Structurize to use the same choice as MineColonies.
Specifically, the initial anchor is the block adjacent to the clicked face of the clicked block. For the shape tool, this lets you build against a surface nicely. For the build tool, the assumption is that most of the time you will be clicking the top face of the "ground" block anyway. (And when the hut/anchor block is sitting directly on top of the ground, then this is already the correct location.)
This also adjusts the treatment of schematics without a groundlevel tag to assume that the ground level is one block below the anchor (which agrees with the above assumption, and with the wiki docs for the groundlevel tag). This is not true of all existing schematics, but there is no rule that would work universally.