Structurize 1.5 (#517)
This has the all the adjustments for the new structurize backend.
Could almost be called half way of a structurize 2.0

Changes proposed in this pull request:
Loading structures more flexibly via structure packs
Syncing between client and server
Additional metadata in blueprints
Support to render multiple blueprints at the same time
Async handling of blueprints.
A few details:

The API is the StructurePacks file, where blueprints can be Stored and retrieved from given a structure pack id (which leads to a structure pack) and the path within the structure pack. Structure pack metadata allows to locate the root of the structure pack folder.

The client and server loaders are mostly for client/server synching on startup and to initially discover structure packs. However, mods can also dynamically adjust their packs whenever they want.

Structure packs are stored in the mod resources but neither in the datapacks nor in the assets to avoid confusion. We handle syncing ourselves and treat the packs as "instance global". Similarly, locally players can drop them off in the root instance folder in a "blueprints" folder.

On scan, a special player folder is created inside this blueprints folder for the scans.
Client side scans are also synced in a special way to the server to avoid servers having to distribute all of those to the other clients.