Scan & command tweaks (#554)
Increases the length limit on the scan tool name -- now it can (and should) contain a path, so longer names are needed
Allows the paste and pasteFolder commands to use quoted strings instead of dotted.words.
The dot conversions are still supported for backwards compatibility. But perhaps could be removed for 1.20. (Also why didn't they just use String.replace?)
Gives better error responses for paste and pasteFolder in case the pack and/or blueprint specified don't actually exist.