Chore: new Delegate blueprint iterator implementation (#670)
Adds a new AbstractDelegateBlueprintIterator class, which can be used to create new BlueprintIterators based on other iterators,
to e.g. apply extra logic to incrementing and decrementing, whilst not being dependent on a particular iterator type
The logic in AbstractBlueprintIterator has been separated in an interface (which is implemented in the delegate iterator as well), a new abstract class (to share the iterator with skip condition logic), and the actual implementation
The latter has been renamed into a new class AbstractStructureIterator, whose name has been inspired by the StructureIterators "registry" class. That is now the base class of the different iterator types
This will be used by a rework of the quarrier in Minecolonies, which will use an approach with a "LayerBlueprintIterator", which iterates over a single Y-level and then finishes, using a delegate iterator which defines the actual iteration pattern for that layer